SectionFiller: Rename file to SectionFilling
This is not an object anymore.
